src/main/java/upside/federation/FederatedCategory.java
/*
 * Copyright 2010 Torkjel Hongve (torkjelh@conduct.no)
 *
 * This file is part of Upside.
 *
 * Upside is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * Upside is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with Upside.  If not, see <http://www.gnu.org/licenses/>.
 */
package upside.federation;

import java.util.HashSet;
import java.util.Set;

class FederatedCategory {
    private String name;
    private String description;
    private Set<Include> includes;

    public FederatedCategory(String name, String description, Set<Include> includes) {
        this.name = name;
        this.description = description;
        this.includes = new HashSet<Include>(includes);
    }

    public String getName() {
        return name;
    }

    public String getDescription() {
        return description;
    }

    public Set<Include> getIncludes() {
        return new HashSet<Include>(includes);
    }
}
